requirements.txt
discord.py
mcstatus
python-dotenv